MiniCalendar is now a view
